render_writeback: the copy into guest pages moves below the settle it is not
`render_writeback.rs` held two rails' worth of work under one roof. The
settle half is neutral — the sites, and the wait a host reader of guest
bytes owes a rail that may still be writing them — and it stays neutral by
asking `Backend::guest_writes_outstanding` for the answer. The copy half is
Vulkan's alone, and says so in every signature: each entry point takes a
`TargetIdentity`, and a rail that holds no residents has nothing to name.

Eighteen `cfg(feature = "backend-vulkan")` items sat below the settle,
contiguous but unmarked. They become `render_writeback/vulkan.rs`, gated
once on the `mod`; the file that remains carries no rail's name in any
signature. Callers — all of them already inside the Vulkan rail — name
`render_writeback::vulkan::…`, which is what they were reaching for.

Pure placement; no item's body changed. metal 1397, vulkan 2181, both 2265,
all unchanged; matrix, fmt and clippy clean, no new rustdoc warnings.
